# sdk/go/api/openapi.yaml
openapi: 3.1.0
info:
  description: "C2PA-compliant content signing and verification infrastructure for\
    \ publishers, legal/finance firms, AI labs, and enterprises"
  title: Encypher Enterprise API
  version: 1.0.0-preview
servers:
- description: Production
  url: https://api.encypherai.com
- description: Local development
  url: http://localhost:8007
paths:
...
  /api/v1/documents:
    get:
      description: |-
        List signed documents for the organization.

        Supports pagination, search, and filtering by status and date range.
      operationId: list_documents_api_v1_documents_get
      parameters:
      - description: Page number
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          description: Page number
          minimum: 1
          title: Page
          type: integer
        style: form
      - description: Items per page
        explode: true
        in: query
        name: page_size
        required: false
        schema:
          default: 50
          description: Items per page
          maximum: 100
          minimum: 1
          title: Page Size
          type: integer
        style: form
      - description: Search in title
        explode: true
        in: query
        name: search
        required: false
        schema:
          nullable: true
          type: string
        style: form
      - description: Filter by status (active/revoked)
        explode: true
        in: query
        name: status
        required: false
        schema:
          nullable: true
          type: string
        style: form
      - description: Filter from date (ISO format)
        explode: true
        in: query
        name: from_date
        required: false
        schema:
          nullable: true
          type: string
        style: form
      - description: Filter to date (ISO format)
        explode: true
        in: query
        name: to_date
        required: false
        schema:
          nullable: true
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: List Documents
      tags:
      - Documents
      - Documents
  /api/v1/documents/{document_id}:
    delete:
      description: |-
        Soft delete a document.

        By default, this also revokes the document so it will fail verification.
        The document is not permanently deleted but marked as deleted.
      operationId: delete_document_api_v1_documents__document_id__delete
      parameters:
      - explode: false
        in: path
        name: document_id
        required: true
        schema:
          title: Document Id
          type: string
        style: simple
      - description: Also revoke the document
        explode: true
        in: query
        name: revoke
        required: false
        schema:
          default: true
          description: Also revoke the document
          title: Revoke
          type: boolean
        style: form
      - description: Reason for deletion
        explode: true
        in: query
        name: reason
        required: false
        schema:
          nullable: true
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDeleteResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Delete Document
      tags:
      - Documents
      - Documents
    get:
      description: Get detailed information about a specific document.
      operationId: get_document_api_v1_documents__document_id__get
      parameters:
      - explode: false
        in: path
        name: document_id
        required: true
        schema:
          title: Document Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetailResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get Document
      tags:
      - Documents
      - Documents
  /api/v1/documents/{document_id}/history:
    get:
      description: |-
        Get the audit trail/history for a document.

        Shows all actions taken on the document including signing, verification, and revocation.
      operationId: get_document_history_api_v1_documents__document_id__history_get
      parameters:
      - explode: false
        in: path
        name: document_id
        required: true
        schema:
          title: Document Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentHistoryResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get Document History
      tags:
      - Documents
      - Documents
...
  /api/v1/status/documents/{document_id}:
    get:
      description: Get the revocation status of a document.
      operationId: get_document_status_api_v1_status_documents__document_id__get
      parameters:
      - explode: false
        in: path
        name: document_id
        required: true
        schema:
          title: Document Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentStatusResponse"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get Document Status
      tags:
      - Status & Revocation
      - Status & Revocation
...
components:
  schemas:
...
    DocumentListResponse:
      description: Response for document listing.
      example:
        data:
          key: ""
        success: true
      properties:
        success:
          default: true
          title: Success
          type: boolean
        data:
          additionalProperties: {}
          description: Documents and pagination info
          title: Data
      required:
      - data
      title: DocumentListResponse
...
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          type: type
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          type: type
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
      title: HTTPValidationError
...
    DocumentHistoryResponse:
      description: Response for document history.
      example:
        data:
          key: ""
        success: true
      properties:
        success:
          default: true
          title: Success
          type: boolean
        data:
          additionalProperties: {}
          title: Data
      required:
      - data
      title: DocumentHistoryResponse
...
    DocumentDeleteResponse:
      description: Response for document deletion.
      example:
        data:
          key: ""
        success: true
      properties:
        success:
          default: true
          title: Success
          type: boolean
        data:
          additionalProperties: {}
          title: Data
      required:
      - data
      title: DocumentDeleteResponse
...
    ValidationError:
      example:
        msg: msg
        loc:
        - ValidationError_loc_inner
        - ValidationError_loc_inner
        type: type
      properties:
        loc:
          items:
            $ref: "#/components/schemas/ValidationError_loc_inner"
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    DocumentDetailResponse:
      description: Response for single document.
      example:
        data:
          word_count: 0
          revoked_at: revoked_at
          signer_id: signer_id
          signed_at: signed_at
          revoked_reason: revoked_reason
          document_id: document_id
          title: title
          url: url
          document_type: document_type
          status: active
          verification_url: verification_url
        success: true
      properties:
        success:
          default: true
          title: Success
          type: boolean
        data:
          $ref: "#/components/schemas/DocumentDetail"
      required:
      - data
      title: DocumentDetailResponse
    DocumentDetail:
      description: Detailed document information.
      example:
        word_count: 0
        revoked_at: revoked_at
        signer_id: signer_id
        signed_at: signed_at
        revoked_reason: revoked_reason
        document_id: document_id
        title: title
        url: url
        document_type: document_type
        status: active
        verification_url: verification_url
      properties:
        document_id:
          title: Document Id
          type: string
        title:
          nullable: true
          type: string
        document_type:
          nullable: true
          type: string
        status:
          default: active
          title: Status
          type: string
        signed_at:
          title: Signed At
          type: string
        verification_url:
          title: Verification Url
          type: string
        word_count:
          nullable: true
          type: integer
        url:
          nullable: true
          type: string
        signer_id:
          nullable: true
          type: string
        revoked_at:
          nullable: true
          type: string
        revoked_reason:
          nullable: true
          type: string
      required:
      - document_id
      - signed_at
      - verification_url
      title: DocumentDetail
...
    DocumentStatusResponse:
      description: Response for document status query.
      example:
        status_bit_index: 6
        revoked_at: revoked_at
        organization_id: organization_id
        revoked_reason_detail: revoked_reason_detail
        status_list_index: 0
        reinstated_at: reinstated_at
        revoked_reason: revoked_reason
        document_id: document_id
        revoked: true
        status_list_url: status_list_url
      properties:
        document_id:
          title: Document Id
          type: string
        organization_id:
          title: Organization Id
          type: string
        revoked:
          title: Revoked
          type: boolean
        revoked_at:
          nullable: true
          type: string
        revoked_reason:
          nullable: true
          type: string
        revoked_reason_detail:
          nullable: true
          type: string
        reinstated_at:
          nullable: true
          type: string
        status_list_url:
          nullable: true
          type: string
        status_list_index:
          nullable: true
          type: integer
        status_bit_index:
          nullable: true
          type: integer
      required:
      - document_id
      - organization_id
      - revoked
      title: DocumentStatusResponse
...
    ValidationError_loc_inner:
      anyOf:
      - type: string
      - type: integer
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http